+++ examples/flutter_main/lib/infinite_list/infinite_list.dart
@@ -1,0 +1,11 @@
+import 'dart:math' as math;
+
+import 'package:example/const.dart';
+import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+import 'package:state_beacon/state_beacon.dart';
+
+part 'controller.dart';
+part 'models.dart';
+part 'repo.dart';
+part 'view.dart';
