+++ examples/flutter_main/lib/main.dart
@@ -1,7 +1,7 @@
 // ignore_for_file: prefer_const_constructors, prefer_const_literals_to_create_immutables
 
 import 'package:example/counter_page.dart';
-import 'package:example/inifite_list.dart';
+import 'package:example/infinite_list/infinite_list.dart';
 import 'package:example/konami_page.dart';
 import 'package:example/search_page.dart';
 import 'package:example/todo_page.dart';
